requirements.txt
# Core Dependencies
numpy>=1.24.0
scipy>=1.10.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Deep Learning
torch>=2.0.0

# Sionna (Wireless Communication Simulator)
sionna>=0.18.0

# Reinforcement Learning
gymnasium>=0.29.0
stable-baselines3>=2.1.0
sb3-contrib>=2.1.0

# Utilities
pyyaml>=6.0
tqdm>=4.65.0
pandas>=2.0.0
tensorboard>=2.15.0

# Jupyter for analysis
jupyter>=1.0.0
ipykernel>=6.25.0
ipywidgets>=8.1.0

# Plotting
plotly>=5.17.0

# For saving/loading models
cloudpickle>=2.2.0

